Downgrade socket.io types to same version that nodecg uses
In version `1.8.1` of our typings you will get an error if you want to use a `NodeCG` instance from these typings somewhere where a `NodeCG` instance of the official typings is expected. This happend because the types of the returned socker.io servers (from `NodeCGServer.getSocketIOServer()`) would differ.
Before this PR this package uses the `3.x` series which are just stubs because with these versions socket.io bundles their own definitions.
To fix this and get the same typings we need to depend on the same socket.io typings from version 2.x/1.x.

Will be released as version `1.8.2`.